refactor(backend): OrganizationContactPersonNotesCreateController: move functionality of organization creation with associated contact person and notes to a more specific controller
