--- app/Http/Controllers/NotesController.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\Notes;
-use Illuminate\Http\Request;
-
-class NotesController extends Controller
-{
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        // storing is happening in the OrganizationController
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, Notes $notes)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Notes $notes)
-    {
-        //
-    }
-}
